// frontend/types.ts
export interface ResearchNote {
    id: number;
    title: string;
    content: string;
    last_updated: string;
}